fix: update based on feedbacks
